requirement.txt
python-telegram-bot[ext]==20.0
telethon==1.34.0
async-lru==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pyaes==1.6.1
pycryptodome==3.18.0
pyrogram==2.0.106
pyromod==1.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
TgCrypto==1.2.5
urllib3==2.0.3
websockets==11.0.3
yt-dlp==2023.6.22
motor==3.1.2
aiohttp==3.8.4
aiofiles==23.1.0
pytz==2023.3
ffmpeg-python==0.2.0
umongo==3.1.0
speedtest-cli==2.1.3
Flask==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
werkzeug==2.0.2
itsdangerous==2.0.1
m3u8==3.5.0
pycryptodomex==3.18.0
uvloop==0.17.0
aiosqlite==0.18.0